refactor(server): extract handleGetMatchSummary to dedicated module (S27.8.32)
Final polish slice. Bring `match-details-handlers.ts` under DoD
secondary 400 by extracting `handleGetMatchSummary` (~104 lines)
into `routes/match-summary-handler.ts` (142 lines).

Endpoint covered:
- `GET /:id/summary` — handleGetMatchSummary (prematch summary:
  teams, coaches, score approx, turn/half, acceptances)

Re-export chain preserved (`match.ts` -> `match-details-handlers.ts`
-> handler).

`match-details-handlers.ts` shrinks 464 -> 358 lines (under DoD
secondary 400).

**All extracted modules are now under DoD secondary 400 lines.**
The polish phase of the S27.8 refactor is complete.

4 new smoke tests `match-summary-handler.test.ts`:
- exports x3 (named + 2 re-exports through chain)
- 404 gate on match not found

- 1850/1850 server tests pass
- typecheck clean
